// Nebura Client/src/typings/discord.d.ts
export interface componentData {
  /**
   * Unique identifier for the component.
   * This ID is used to distinguish the component from others.
   */
  id: string;

  /**
   * Indicates if the component is related to tickets.
   * If `true`, the component is associated with ticketing functionality.
   */
  tickets: boolean;

  /**
   * Indicates if the component is restricted to the bot owner.
   * If `true`, only the bot owner can interact with this component.
   */
  owner: boolean;

  /**
   * Indicates if the component is under maintenance.
   * If `true`, the component is temporarily disabled for updates or fixes.
   */
  maintenance?: boolean;

  /**
   * Permissions required by the user to interact with the component.
   * This is an array of Discord permissions that the user must have.
   */
  permissions: PermissionResolvable[];

  /**
   * Permissions required by the bot to execute the component.
   * This is an array of Discord permissions that the bot must have.
   */
  botpermissions: PermissionResolvable[];
}

/**
 * Interface for button components.
 * This interface extends `componentData` and adds functionality specific to buttons.
 */
export interface Buttons extends componentData {
  /**
   * Function to execute when the button is interacted with.
   *
   * @param interaction - The button interaction object provided by Discord.js.
   * @param client - The bot client instance, which manages the bot's state and interactions.
   * @param language - The language code for localization, used to provide responses in the user's language.
   * @param configuration - The bot's configuration object, containing global settings.
   */
  execute: (
    interaction: ButtonInteraction,
    client: MyClient,
    language: string,
    configuration: typeof config,
  ) => void;
}

/**
 * Interface for menu components.
 * This interface extends `componentData` and adds functionality specific to menus.
 */
export interface Menus extends componentData {
  /**
   * Function to execute when the menu is interacted with.
   *
   * @param interaction - The menu interaction object, which can be one of several types:
   * `StringSelectMenuInteraction`, `ChannelSelectMenuInteraction`, or `RoleSelectMenuInteraction`.
   * @param client - The bot client instance, which manages the bot's state and interactions.
   * @param language - The language code for localization, used to provide responses in the user's language.
   * @param configuration - The bot's configuration object, containing global settings.
   */
  execute: (
    interaction:
      | StringSelectMenuInteraction
      | ChannelSelectMenuInteraction
      | RoleSelectMenuInteraction,
    client: MyClient,
    language: string,
    configuration: typeof config,
  ) => void;
}

/**
 * Interface for modal components.
 * This interface extends `componentData` and adds functionality specific to modals.
 */
export interface Modals extends componentData {
  /**
   * Function to execute when the modal is submitted.
   *
   * @param interaction - The modal submit interaction object provided by Discord.js.
   * @param client - The bot client instance, which manages the bot's state and interactions.
   * @param language - The language code for localization, used to provide responses in the user's language.
   * @param configuration - The bot's configuration object, containing global settings.
   */
  execute: (
    interaction: ModalSubmitInteraction,
    client: MyClient,
    language: string,
    configuration: typeof config,
  ) => void;
}